generate initial commit
