+++ stl/inc/algorithm
@@ -1261,43 +1261,6 @@ namespace ranges {
 
     inline constexpr _Copy_if_fn copy_if{_Not_quite_object::_Construct_tag{}};
 } // namespace ranges
-
-namespace ranges {
-    // ALIAS TEMPLATE move_result
-    template <class _In, class _Out>
-    using move_result = in_out_result<_In, _Out>;
-
-    // VARIABLE ranges::move
-    class _Move_fn : private _Not_quite_object {
-    public:
-        using _Not_quite_object::_Not_quite_object;
-
-        // clang-format off
-        template <input_iterator _It, sentinel_for<_It> _Se, weakly_incrementable _Out>
-            requires indirectly_movable<_It, _Out> 
-        constexpr move_result<_It, _Out> operator()(_It _First, _Se _Last, _Out _Result) const {
-            _Adl_verify_range(_First, _Last);
-            auto _UFirst      = _Get_unwrapped(_STD move(_First));
-            const auto _ULast = _Get_unwrapped(_STD move(_Last));
-            for (; _UFirst != _ULast; ++_UFirst, (void) ++_Result) {
-                *_Result = _RANGES iter_move(_UFirst);
-            }
-
-            _Seek_wrapped(_First, _STD move(_UFirst));
-            return {_STD move(_First), _STD move(_Result)};
-        }
-
-        template <input_range _Rng, weakly_incrementable _Out>
-            requires indirectly_movable<iterator_t<_Rng>, _Out> 
-        constexpr move_result<borrowed_iterator_t<_Rng>, _Out>
-            operator()(_Rng&& _Range, _Out _Result) const {
-            return (*this)(_RANGES begin(_Range), _RANGES end(_Range), _STD move(_Result));
-        }
-        // clang-format on
-    };
-
-    inline constexpr _Move_fn move{_Not_quite_object::_Construct_tag{}};
-} // namespace ranges
 #endif // __cpp_lib_concepts
 
 // FUNCTION TEMPLATE partition_copy
@@ -2384,61 +2347,6 @@ _FwdIt2 replace_copy_if(_ExPo&&, _FwdIt1 _First, _FwdIt1 _Last, _FwdIt2 _Dest, _
     return _STD replace_copy_if(_First, _Last, _Dest, _Pass_fn(_Pred), _Val);
 }
 #endif // _HAS_CXX17
-
-#ifdef __cpp_lib_concepts
-namespace ranges {
-
-    // VARIABLE ranges::fill
-    class _Fill_fn : private _Not_quite_object {
-    public:
-        using _Not_quite_object::_Not_quite_object;
-
-        // clang-format off
-            template <class _T, output_iterator<const _T&> _O, sentinel_for<_O> _S>
-            constexpr _O operator()(_O _First, _S _Last, const _T& _Value) const {
-                _Adl_verify_range(_First, _Last);
-                auto _UFirst = _Get_unwrapped(_STD move(_First));
-                auto _ULast = _Get_unwrapped(_STD move(_Last));
-                while (_UFirst != _ULast) {
-                    *_UFirst = _Value;
-                    ++_UFirst;
-                }
-
-                _Seek_wrapped(_First, _STD move(_UFirst));
-                return _First;
-            }
-
-            template <class _T, output_range<const _T&> _Rng>
-            constexpr borrowed_iterator_t<_Rng> operator()(_Rng&& _Range, const _T& _Value) const {
-                return (*this)(_RANGES begin(_Range), _RANGES end(_Range), _Value);
-            }
-        // clang-format on
-    };
-
-    inline constexpr _Fill_fn fill{_Not_quite_object::_Construct_tag{}};
-
-    // VARIABLE ranges::fill_n
-    class _Fill_n_fn : private _Not_quite_object {
-    public:
-        using _Not_quite_object::_Not_quite_object;
-
-        // clang-format off
-            template <class _T, output_iterator<const _T&> _O>
-            constexpr _O operator()(_O _First, iter_difference_t<_O> _Count, const _T& _Value) const {
-                auto _UFirst = _Get_unwrapped_n(_STD move(_First), _Count);
-                for (; _Count > 0; ++_UFirst, --_Count) {
-                    *_UFirst = _Value;
-                }
-
-                _Seek_wrapped(_First, _STD move(_UFirst));
-                return _First;
-            }
-        // clang-format on
-    };
-
-    inline constexpr _Fill_n_fn fill_n{_Not_quite_object::_Construct_tag{}};
-} // namespace ranges
-#endif // __cpp_lib_concepts
 
 #ifdef __cpp_lib_concepts
 namespace ranges {
