+++ tests/std/tests/P0896R4_ranges_alg_generate/env.lst
